requirements.txt
fastapi
docker
uvicorn
fastapi~=0.68.1
sqlalchemy~=1.4.23
bcrypt~=3.2.0
pyjwt~=2.1.0
python-dotenv~=0.19.0
setuptools~=52.0.0
boto3